qemu: Enable security monitor and virtualization
--------

QEMU's Cortex-A53 & Cortex-A72 emulation by default only emulates a CPU with EL1(CurrentEL==4),
not EL2 or EL3, and therefore (as the architecture requires) we boot up in EL1.
If you want a CPU with EL2 or EL3 you need to specifically
enable it on the command line, with
-machine virtualization=on to enable EL2,
and -machine secure=on to enable EL3.

Signed-off-by: Xingfeng Yang <nerosoft@outlook.com>